Project/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
tortoise-orm==0.25.0
asyncpg==0.30.0
slowapi==0.1.9